показывает путь на карте, populate database не работает, не запрашивает права

+++ app/src/main/java/ru/vanchikov/fitnesinmylife/data/DAO/UsersWaysDao.kt
@@ -31,7 +31,7 @@ interface UsersWaysDao {
     fun allUserWaysByUserId(userId : String) : LiveData<List<UserWays>>
 
     @Query("SELECT * FROM wayfix_table where wayId = :wayId")
-    fun allWayFixByWayId(wayId: Int) : LiveData<List<WayFix>>
+    fun allWayFixByWayId(wayId: Long) : LiveData<List<WayFix>>
 
     @Query("SELECT * FROM users_ways_table where wayId = :wayId")
     fun getWayById(wayId: Long) : UserWays
